Develop

+++ app/src/main/java/github/detrig/weatherapp/findcity/domain/FoundCity.kt
@@ -5,6 +5,6 @@ import java.io.Serializable
 data class FoundCity(
     val name: String,
     val country: String,
-    val latitude: Double,
-    val longitude: Double
+    val latitude: Float,
+    val longitude: Float
 ) : Serializable
